docker/pong/requirements.txt
Django==3.2.5
psycopg2==2.8.6
djangorestframework==3.12.4
requests==2.32.3
djangorestframework_simplejwt==5.3.1
django-cors-headers==4.4.0

daphne==3.0.2
channels==3.0.4
channels_redis==3.3.1